Refactor: split files inclusion loop from files finding loop
This is necessary, since in https://github.com/doctrine/persistence/pull/429 we would like to use a different way of "finding files" - using a client-defined iterable. Thus, the only way to do the switch is to split the loops.
