+++ tests/Persistence/Mapping/_files/colocated/Entity.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Doctrine\Tests\Persistence\Mapping\_files\colocated;
+
+/**
+ * The driver should include this file and return its class name
+ * from {@see \Doctrine\Persistence\Mapping\Driver\ColocatedMappingDriver::getAllClassNames()} method.
+ */
+class Entity
+{
+}
