+++ tests/Persistence/Mapping/_files/colocated/EntityFixture.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Doctrine\Tests\Persistence\Mapping\_files\colocated;
+
+/**
+ * This class is in the same directory as {@see Entity},
+ * therefore, it's included when using an ordinary path scanning mechanism.
+ */
+class EntityFixture
+{
+}
