+++ tests/Persistence/Mapping/_files/colocated/Foo.mphp
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Doctrine\Tests\Persistence\Mapping\_files\colocated;
+
+/**
+ * This file has a .mphp extension, and its presence verifies
+ * that preg_quote() is actually called on the file extension.
+ * Thus, it never reaches the driver as it only uses .php extension.
+ */
+final class Foo
+{
+}
