+++ tests/Persistence/Mapping/_files/colocated/TestClass.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Doctrine;
+namespace Doctrine\Tests\Persistence\Mapping\_files\colocated;
 
+/**
+ * This class is considered transient {@see \Doctrine\Persistence\Mapping\Driver\ColocatedMappingDriver::isTransient()}
+ * by the driver, and therefore its class name is not returned.
+ */
 class TestClass
-{
-}
-
-class Entity
 {
 }
